added section dom, moved handlers to own file
